After Project submission commit
Made improvements to the README file and made a BAT file to allow for the program to run on Windows.

Added a number of comments and other things to attempt to aid in understanding the program. Also added the resign functionality to deal with players leaving a multiplayer session.

Dealt with weird indentation and fixed issue with font and turnFill